backend/requirements.txt
tensorflow 
flask
flask_cors
bcrypt
python-dotenv 
opencv-python
scikit-learn
flask-socketio
imutils
pymongo
PyJWT